Fixed significant protocol handling error around timers
Two separate problems combined to create corrupted frames commonly and
potentially starving new clients.  First, NewClientClearToSend was
incorrectly marked as not expecting a reply but in reality it must
expect one (even if one never arrives) to give space for the titular new
client to respond.  Second, when timer ticks are "skipped" due to an
existing authorized sender this causes non-deterministic behaviour
because the timer tick count is incremented but the behaviour is
skipped.  This in practice could cause, for example, StatusUpdate
messages to never get sent as we are always skipping the 2nd tick after
a NewClientClearToSend.  To combat this, we can relax the notion that
the timer is actually about timing and instead just run each tick when
we can, then increment.  This means that clients will observe variable
delays in receiving a reply from the main board but this shouldn't
actually matter in practice as they have multiple attempts each second.